Convert cell to implicitly unwrapped optional
Added a little explanation about why in the post, but essentially if we're immediately checking `cell` against `nil`, there's no harm in declaring it as implicitly unwrapped -- it even makes sense, since implicitly unwrapped optionals are optionals that you're basically promising won't be `nil`. I especially like it this way better since then we aren't using `!` over and over in the code below, training ourselves out of aversion to the force-unwrapping operator.